requirements.txt
# Core ML and Audio Processing
torch
sentence-transformers
openai-whisper

# Utility & Translation
gTTS
deep-translator
requests
langdetect==1.0.9

# Backend & Server Framework
fastapi
uvicorn[standard]

# Optional Frontend
streamlit

# Multipart handling
python-multipart
